% Generated by roxygen2 (4.0.2): do not edit by hand
\name{AvoidRedundancy}
\alias{AvoidRedundancy}
\title{Cleans data to avoid redundancy}
\usage{
AvoidRedundancy(x, newitem)
}
\arguments{
\item{x}{A data frame.}

\item{newitem}{Variable for which to check}
}
\value{
A dataframe without data on newitem
}
\description{
Checks if a variable is already present in a dataset and removes it. This is
useful if you use a function like ProductNR or RatioNR, where often a
certain variable is available for historical, but not for scenario data
}
\author{
Niklas Roming
}

