% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BestBalancedPanelNR}
\alias{BestBalancedPanelNR}
\title{Balanced panel with the highest number of observations}
\usage{
BestBalancedPanelNR(data, myitem)
}
\arguments{
\item{data}{NRmisc compatible data frame.}

\item{myitem}{Variable for which to find the best balanced panel.}
}
\value{
NRmisc compatible data frame containing the best balanced panel.
}
\description{
Balanced panel with the highest number of observations
}
\author{
Niklas Roming
}

