% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrmisc.R
\docType{data}
\name{cgc_gene_names}
\alias{cgc_gene_names}
\title{Cancer gene census gene names}
\format{character vector}
\usage{
cgc_gene_names
}
\description{
COSMIC (v80) cancer gene census, as downloaded Feb 20 2017
}
\keyword{datasets}
