% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour.R
\name{col_show}
\alias{col_show}
\title{Plot colour palette}
\usage{
col_show(cols)
}
\arguments{
\item{cols}{Vector of colours}
}
\description{
Plot colour palette
}
\examples{
mycol <- nr_col_pal("colblind", 9)
col_show(mycol)
}
\seealso{
\code{\link{nr_col_pal}}
}

