% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_tools.R
\name{collapse_granges_by_val}
\alias{collapse_granges_by_val}
\title{Collapse consecutive genomic ranges with same meta-data value}
\usage{
collapse_granges_by_val(gr, na.rm = TRUE)
}
\arguments{
\item{gr}{A GRanges object with one meta-data column}

\item{na.rm}{Logical - if TRUE (default), ranges with NA in metadata column are removed from output.
If FALSE, ranges with NA are retained AND gaps between ranges are returned (also with NA)}
}
\value{
A GRanges object where consecutive ranges with the same meta-data
 value are collapsed together
}
\description{
Collapse consecutive genomic ranges with same meta-data value
}
\examples{
library(GenomicRanges)
my_gr <- GRanges(c("chr1", "chr1", "chr1", "chr1", "chr2"),
                 IRanges(start=c(10, 20, 40, 50, 60),
                         end=c(19, 29, 49, 59, 69)),
                 value=c(1, 2, 2, 2, 2))
collapse_granges_by_val(my_gr)
}
