% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ContinueGrowthNR}
\alias{ContinueGrowthNR}
\title{Lets historical growth continue}
\usage{
ContinueGrowthNR(x)
}
\arguments{
\item{x}{Data frame containing a column 'year', unique in 'item' and 'iso3c'}
}
\description{
Lets historical growth continue
}
\author{
Niklas Roming
}

