% Generated by roxygen2 (4.0.2): do not edit by hand
\name{FadeOutGrowthNR}
\alias{FadeOutGrowthNR}
\title{Continues a time series by fading out historical change rates over a certain
time}
\usage{
FadeOutGrowthNR(tmp, fade.time)
}
\arguments{
\item{tmp}{A NRmisc compatible data frame unique wrt. 'iso3c', 'scenario' and 'item'.}
}
\description{
Continues a time series by fading out historical change rates over a certain
time
}
\author{
Niklas Roming
}

