% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GetGrowthRateNR}
\alias{GetGrowthRateNR}
\title{Computes average growth rate of a time series}
\usage{
GetGrowthRateNR(x)
}
\arguments{
\item{x}{Data frame containing a column 'year', unique in 'item' and 'iso3c'}
}
\value{
Average growth rate of 'value' column in the dataframe
}
\description{
Computes average growth rate of a time series
}
\author{
Niklas Roming
}

