% Generated by roxygen2 (4.0.2): do not edit by hand
\name{GrowthNR}
\alias{GrowthNR}
\title{Lets historical data grow}
\usage{
GrowthNR(data, myitem, hist.time, scen.time, fade.time)
}
\arguments{
\item{data}{NRmis compatible data frame}

\item{myitem}{Item/variable for which calculations will be made.}

\item{hist.time}{Historical periods to take into account (e.g. for
calculation of growth rates)}

\item{scen.time}{Scenario perios to take into account (for which new values
will be calculated)}

\item{fade.time}{Number of periods over which to fade out historical growth
rates}
}
\description{
Lets historical data grow
}
\author{
Niklas Roming
}

