% Generated by roxygen2 (4.0.2): do not edit by hand
\name{IntermediateTimeStepsNR}
\alias{IntermediateTimeStepsNR}
\title{Computes values for intermediate time steps by linear interpolation}
\usage{
IntermediateTimeStepsNR(data, myitem, ts.length = 5, hist.scen, temp.hist)
}
\arguments{
\item{data}{NRmisc compatible data frame}

\item{myitem}{Variable for which intermediate time steps are to be computed}

\item{hist.scen}{Scenario name of historical data.}

\item{temp.hist}{Years of historical data (last time step of historical data
is needed for computing values between historical data and first scenario
data)}

\item{ts.lenght}{Target time step length}
}
\value{
NRmisc data frame with the new time steps attached to the original
data
}
\description{
Computes values for intermediate time steps by linear interpolation
}
\author{
Niklas Roming
}

