% Generated by roxygen2 (4.0.2): do not edit by hand
\name{InterpolateNR}
\alias{InterpolateNR}
\title{Interpolates (linearly) values for gaps in time-series}
\usage{
InterpolateNR(data, myitem, zeroAsNA=TRUE)
}
\arguments{
\item{data}{A NRmisc compatible data frame}

\item{myitem}{Variable for which interpolation shall be carried out.}

\item{zerosAsNA}{Treat zeros as missing values or not?}
}
\value{
A NRmisc compatible data frame without the gaps in the respective
time-series
}
\description{
Interpolates (linearly) values for gaps in time-series
}
\author{
Niklas Roming
}

