% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LargeNR}
\alias{LargeNR}
\title{Return n-largest value of a vector of numbers}
\usage{
LargeNR(x, n)
}
\arguments{
\item{x}{A vector (numeric or integer)}

\item{n}{Which value to return (e.g. n=2 => second-largest)}
}
\value{
The n-largest value of x
}
\description{
Return n-largest value of a vector of numbers
}
\author{
Niklas Roming
}

