% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LMIsoNR}
\alias{LMIsoNR}
\title{Return countries for which LSDV estimation was carried out}
\usage{
LMIsoNR(lm, formula)
}
\arguments{
\item{lm}{A linear model}

\item{formula}{The formula used for estimation (TODO: explain why this is needed)}
}
\value{
A character vector containing the ISO codes of the country fixed
effects
}
\description{
Return countries for which LSDV estimation was carried out
}
\author{
Niklas Roming
}

