% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_tools.R
\name{plot_gr_track}
\alias{plot_gr_track}
\title{Plot one quantitative meta-data track from a GRanges object}
\usage{
plot_gr_track(gr, track, seqname, start = NULL, end = NULL)
}
\arguments{
\item{gr}{A GRanges object}

\item{track}{Name of quantitative meta-data track to plot}

\item{seqname}{seqname to plot for}

\item{start}{starting position of plot (start of seqname if left NULL)}

\item{end}{end position of plot (end of seqname if left NULL)}
}
\description{
Plot one quantitative meta-data track from a GRanges object
}
