% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ProdNR}
\alias{ProdNR}
\title{Calculates the product of two variables}
\usage{
ProdNR(data, newitem, factor1, factor2, redundant = FALSE)
}
\arguments{
\item{data}{A data frame in standard long format.}

\item{newitem}{Name of the new variable that is calculated.}

\item{factor1}{Name of the first variable to be multiplied.}

\item{factor2}{Name of the second variable to be multiplied.}

\item{redundant}{Determine whether redundant information should be
included (default = FALSE).}
}
\value{
The original data frame with the newly calculated data appended.
}
\description{
Calculates the product of two variables
}
\author{
Niklas Roming
}

