% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RatioNR}
\alias{RatioNR}
\title{Calculates the ratio of two variables.}
\usage{
RatioNR(data, newitem, numerator, denominator)
}
\arguments{
\item{data}{A data frame in standard long format.}

\item{newitem}{Name of the new variable that is calculated.}

\item{numerator}{Name of the numerator.}

\item{denominator}{Name of the denominator.}
}
\value{
The original data frame with the newly calculated data appended.
}
\description{
Calculates the ratio of two variables.
}
\author{
Niklas Roming
}

