% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_remind_csv}
\alias{read_remind_csv}
\title{Reads a REMIND csv, separated by semicolon}
\usage{
read_remind_csv(path)
}
\arguments{
\item{path}{Path to csv}
}
\value{
A NRmisc compatible data frame
}
\description{
Reads a REMIND csv, separated by semicolon
}
\author{
Niklas roming
}

