% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Refactor factor variables in a data.frame}
\usage{
refactor(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
A data.frame with each factor variable re-factored (old unused levels are dropped)
}
\description{
Refactor factor variables in a data.frame
}
