% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_tools.R
\name{sample_GRanges}
\alias{sample_GRanges}
\title{Sample random positions from a GRanges object}
\usage{
sample_GRanges(gr, n)
}
\arguments{
\item{gr}{A GRanges object}

\item{n}{Number of positions to sample}
}
\value{
A GRanges object of n positions uniformly sampled from gr
}
\description{
Sample random positions from a GRanges object
}
