% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ScenContNR}
\alias{ScenContNR}
\title{Continues historical growth for several scnearios}
\usage{
ScenContNR(data, myitem, scen.time, scenarios)
}
\arguments{
\item{data}{NRmisc compatible data frame.}

\item{myitem}{Used variable.}

\item{scen.time}{Scenario periods.}

\item{scenarios}{Scenarios for which to apply the function.}
}
\description{
Continues historical growth for several scnearios
}
\author{
Niklas Roming
}

