% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ScenFixNR}
\alias{ScenFixNR}
\title{Sets future/scenario values to a fixed historical value}
\usage{
ScenFixNR(data, myitem, history, scenario)
}
\arguments{
\item{data}{A NRmisc compatible data frame.}

\item{myitem}{Variable for which data should be calculated.}

\item{history}{Period of which value will be used for future values.}

\item{scenario}{Periods for which historical value will be used.}
}
\description{
Sets future/scenario values to a fixed historical value
}
\author{
Niklas Roming
}

