% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SmallNR}
\alias{SmallNR}
\title{Return n-smallest value of a vector of numbers}
\usage{
SmallNR(x, n)
}
\arguments{
\item{x}{A vector (numeric or integer)}

\item{n}{Which value to return (e.g. n=2 => second-smallest)}
}
\value{
The n-smallest value of x
}
\description{
Return n-smallest value of a vector of numbers
}
\author{
Niklas Roming
}

