% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SumNR}
\alias{SumNR}
\title{Calculates the sum of variables}
\usage{
SumNR(data, newitem, add, substract)
}
\arguments{
\item{data}{A data frame in standard long format.}

\item{newitem}{Name of the new item to be calculated.}

\item{add}{Vector of variable names to be summed up.}

\item{substract}{Vector of variable names to be substracted.}
}
\value{
The original data frame with the newly calculated data appended.
}
\description{
Calculates the sum of variables
}
\author{
Niklas Roming
}

