% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TopCountries}
\alias{TopCountries}
\title{Identifiers of largest spatial units}
\usage{
TopCountries(x, var, t, n = 10)
}
\arguments{
\item{x}{A data frame.}

\item{var}{The variable or item for which highest realizations shall be taken into account.}

\item{t}{The year for which the maximum values shall be returned.}

\item{n}{The number of spatial units to be returned.}
}
\value{
A character vector (of length n).
}
\description{
Returns identifiers of the n spatial units that have highest values of a
certain variable in a certain year
}
\author{
Niklas Roming
}

