% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_mutations.R
\name{tally_mutations_96}
\alias{tally_mutations_96}
\title{Tally mutations in the 96 base substitution classes defined by trinucleotide context}
\usage{
tally_mutations_96(gr, genome, ref = "ref", alt = "alt",
  samp = "sampleID")
}
\arguments{
\item{gr}{A GRanges object with ranges of width 1 (variant positions)
and at least three metadata columns for reference base, alternate base,
and sample name.}

\item{genome}{A BSgenome object of the reference genome to use for
trinucleotide context calculations (e.g. 'BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19')}

\item{ref}{Variable name for the reference base column in gr (default 'ref').}

\item{alt}{Variable name for the alternative base column in gr (default 'alt').}

\item{samp}{Variable name for the sample ID column in gr (default 'sampleID').}
}
\value{
A matrix of mutation counts with one row per sample and one
 column per base substitution class (trinucleotide context)
}
\description{
Tally mutations in the 96 base substitution classes defined by trinucleotide context
}
