% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tally_mutations.R
\name{tally_mutations_96}
\alias{tally_mutations_96}
\title{Tally mutations in the 96 base substitution classes defined by trinucleotide context}
\usage{
tally_mutations_96(gr)
}
\arguments{
\item{gr}{A GRanges object with ranges of width 1 (variant positions)
and at least three metadata columns called 'ref' for reference base,
'alt' for alternate base, and 'sampleID' for sample name.
Only hg19 chroms chr1-chr22, chrX and chrY are kept.}
}
\value{
A matrix of mutation counts with one row per sample and one
 column per base substitution class (trinucleotide context)
}
\description{
Tally mutations in the 96 base substitution classes defined by trinucleotide context
}

